Address William's review
